+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [9.0.4] - Unreleased
 - Fix bug with duplicate searchables
 - Introduce ability to view Catalogues in a flat view
+- Fix bug where Internal catalogues were still able to be added to a CIC
 
 ## [9.0.3] - 2025-11-03
 - Improve checking for default pipelines